[feat] 로그인 및 토큰 인증 추가, 회원가입 설정, 권한 확인 추가

+++ src/main/java/Project/TravelBusan/controller/UserController.java
@@ -8,6 +8,7 @@ import Project.TravelBusan.request.UserModifyRequestDto;
 import Project.TravelBusan.response.UserListResponseDto;
 import Project.TravelBusan.response.UserLoginResponseDto;
 import Project.TravelBusan.response.ResponseDto;
+import Project.TravelBusan.response.UserResponseDto;
 import Project.TravelBusan.service.UserService;
 import Project.TravelBusan.util.SecurityUtil;
 import jakarta.servlet.http.HttpServletRequest;
@@ -18,6 +19,8 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.prepost.PreAuthorize;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.bind.annotation.*;
 
 import java.io.IOException;
@@ -44,34 +47,46 @@ public class UserController {
     }
 
     @PatchMapping("/{user-id}")
-    public ResponseDto<UserListResponseDto> userModify(@PathVariable("user-id") Long userId, @RequestBody UserModifyRequestDto userModifyRequestDto){
+    public ResponseDto<UserListResponseDto> userModify(@PathVariable("user-id") Long userId, @RequestBody UserModifyRequestDto userModifyRequestDto) throws Exception {
+       if(userModifyRequestDto.getId().equals(userId))
         return userService.updateUserById(userId, userModifyRequestDto);
+       else {
+           throw new IllegalAccessException("본인");
+       }
     }
 
     @DeleteMapping("/{user-id}")
-    public ResponseDto<Void> userRemove(@PathVariable("user-id") Long userId){
-        return userService.deleteById(userId);
+    public ResponseDto<Void> userRemove(@PathVariable("user-id") Long userId, @RequestBody UserResponseDto userResponseDto) throws Exception {
+        if(userResponseDto.getId().equals(userId))
+            return userService.deleteById(userId);
+        else {
+            throw new IllegalAccessException("본인");
+        }
     }
 
-    @PostMapping("/login")
-    public ResponseDto<UserLoginResponseDto> userLogin(@RequestBody UserLoginRequestDto userLoginRequestDto){
-        return userService.login(userLoginRequestDto);
-    }
 
-    /*@PostMapping("/test-redirect")
+
+    @PostMapping("/test-redirect")
     public void testRedirect(HttpServletResponse response) throws IOException {
         response.sendRedirect("/api/user");
-    }*/
+    }
 
     @GetMapping("/user")
-    @PreAuthorize("hasAnyRole('USER','ADMIN')")
     public ResponseEntity<UserLoginRequestDto> getMyUserInfo(HttpServletRequest request) {
         return ResponseEntity.ok(userService.getMyUserWithAuthorities());
     }
 
     @GetMapping("/user/{username}")
-    @PreAuthorize("hasAnyRole('ADMIN')")
-    public ResponseEntity<UserJoinRequestDto> getUserInfo(@PathVariable String username) {
-        return ResponseEntity.ok(userService.getUserWithAuthorities(username));
+    public ResponseEntity<UserLoginRequestDto> getUserInfo(@PathVariable String username, @RequestBody UserResponseDto userResponseDto) throws Exception {
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        if (!hasAdminRole(authentication)) {
+            throw new IllegalAccessException("관리자");
+        } else {
+            return ResponseEntity.ok(userService.getUserWithAuthorities(username));
+        }
+    }
+    private boolean hasAdminRole(Authentication authentication) {
+        return authentication.getAuthorities().stream()
+                .anyMatch(grantedAuthority -> grantedAuthority.getAuthority().equals("ROLE_ADMIN"));
     }
 }
